created evaluate_model.py and moved functions folder under src
