+++ report/diabetes_analysis.html
@@ -124,7 +124,7 @@ div.csl-indent {
 Table&nbsp;1: Accuracy of Logistic Regression and Linear SVC Model
 </figcaption>
 <div aria-describedby="tbl-testing-caption-0ceaefa1-69ba-4598-a22c-09a6ac19f8ca">
-<div class="cell-output cell-output-display cell-output-markdown" data-execution_count="25">
+<div class="cell-output cell-output-display cell-output-markdown" data-execution_count="4">
 <table class="do-not-create-environment cell caption-top table table-sm table-striped small">
 <thead>
 <tr class="header">
@@ -168,7 +168,7 @@ Table&nbsp;1: Accuracy of Logistic Regression and Linear SVC Model
 Table&nbsp;2: Diabetes Data with Health Indicators and Lifestyle Factors
 </figcaption>
 <div aria-describedby="tbl-df-caption-0ceaefa1-69ba-4598-a22c-09a6ac19f8ca">
-<div class="cell-output cell-output-display cell-output-markdown" data-execution_count="26">
+<div class="cell-output cell-output-display cell-output-markdown" data-execution_count="5">
 <table class="do-not-create-environment cell caption-top table table-sm table-striped small">
 <colgroup>
 <col style="width: 2%">
@@ -418,7 +418,7 @@ Figure&nbsp;4: Binary Features by Class
 Table&nbsp;3: Cross Validation Results of Candidate Models
 </figcaption>
 <div aria-describedby="tbl-training-caption-0ceaefa1-69ba-4598-a22c-09a6ac19f8ca">
-<div class="cell-output cell-output-display cell-output-markdown" data-execution_count="27">
+<div class="cell-output cell-output-display cell-output-markdown" data-execution_count="6">
 <table class="do-not-create-environment cell caption-top table table-sm table-striped small">
 <thead>
 <tr class="header">
@@ -431,15 +431,15 @@ Table&nbsp;3: Cross Validation Results of Candidate Models
 <tbody>
 <tr class="odd">
 <td style="text-align: left;">fit_time</td>
-<td style="text-align: right;">0.0069416</td>
-<td style="text-align: right;">0.116905</td>
-<td style="text-align: right;">0.278652</td>
+<td style="text-align: right;">0.0072463</td>
+<td style="text-align: right;">0.117676</td>
+<td style="text-align: right;">0.273912</td>
 </tr>
 <tr class="even">
 <td style="text-align: left;">score_time</td>
-<td style="text-align: right;">0.00148077</td>
-<td style="text-align: right;">0.00418525</td>
-<td style="text-align: right;">0.00422087</td>
+<td style="text-align: right;">0.00149779</td>
+<td style="text-align: right;">0.00444508</td>
+<td style="text-align: right;">0.00413003</td>
 </tr>
 <tr class="odd">
 <td style="text-align: left;">test_score</td>
