+++ results/models/model_training.csv
@@ -1,5 +1,5 @@
 Index,Dummy,Logistic Regression,Linear SVC
-fit_time,0.006941604614257813,0.11690478324890137,0.27865185737609866
-score_time,0.0014807701110839844,0.004185247421264649,0.0042208671569824215
+fit_time,0.0072463035583496095,0.11767606735229492,0.2739123821258545
+score_time,0.001497793197631836,0.004445075988769531,0.004130029678344726
 test_score,0.8609222248366976,0.8637308809265773,0.8635387119695499
 train_score,0.8609222248493607,0.8638392858545858,0.8635461014019216
