+++ src/18_model_testing.py
@@ -12,25 +12,10 @@ from sklearn.svm import LinearSVC
 from sklearn.metrics import accuracy_score
 
 from functions.read_data import read_data
+from functions.evaluate_model import evaluate_model
+from functions.read_model import read_model
 
 # from great_tables import GT
-
-def read_model(out_dir):
-    lr_pipe_path = os.path.join(out_dir, "trained_lr_pipe.pkl")
-    linear_svc_pipe_path = os.path.join(out_dir, "trained_linear_svc_pipe.pkl")
-
-    with open(lr_pipe_path, "rb") as f:
-        lr_pipe = pickle.load(f)
-
-    with open(linear_svc_pipe_path, "rb") as f:
-        linear_svc_pipe = pickle.load(f)
-
-    return lr_pipe, linear_svc_pipe
-
-def evaluate_model(pipe, X_test, y_test):
-    prediction = pipe.predict(X_test)
-    accuracy = accuracy_score(y_test, prediction)
-    return accuracy
 
 
 @click.command()
