--- src/objects/pass_data_validation10.txt
@@ -1,2 +1,0 @@
-result
-PASS
