+++ tests/test_read_model.py
@@ -3,7 +3,7 @@ import os
 
 sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
 
-from src.model_testing import read_model
+from src.functions.read_model import read_model
 
 def test_read_model():
     lr_pipe, linear_svc_pipe = read_model("results/models")
